buttonManager now Owns it's own frame
What point was there to keeping the Frame out in the cold?
How would I explain it in my paragraphs of text?
"Oh yeah, I have this class to hold all the buttons, mentally-wise, which takes as a parameter this outside Frame that holds all the buttons, holding wise". Yari Yari!