Commas in my not-a-function anonymous functions in layout.
I am still learning how to use Gio, I'm afraid.
I'm flitting through the documentation on Layout for laying out the two sections of my App. (The polygon screen, and the buttons)

Why am I committing now? Look at the pattern of commas in the grand (but not working) Background{}.Layout.
This was the only pattern of commas in which I did not get an error. Why? My guess is the comma means "Continued on following lines", something that doesn't make as much sense at the end of a function.